patch: get trackers
